+++ causalpy/pymc_experiments/pre_post_fit.py
@@ -1,0 +1,340 @@
+from typing import Union
+
+import arviz as az
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+import xarray as xr
+from patsy import build_design_matrices, dmatrices
+
+from causalpy.custom_exceptions import BadIndexException  # NOQA
+from causalpy.plot_utils import plot_xY
+from causalpy.pymc_experiments.experimental_design import ExperimentalDesign
+
+LEGEND_FONT_SIZE = 12
+az.style.use("arviz-darkgrid")
+
+
+class PrePostFit(ExperimentalDesign):
+    """
+    A class to analyse quasi-experiments where parameter estimation is based on just
+    the pre-intervention data.
+
+    :param data:
+        A pandas dataframe
+    :param treatment_time:
+        The time when treatment occured, should be in reference to the data index
+    :param formula:
+        A statistical model formula
+    :param model:
+        A PyMC model
+
+    Example
+    --------
+    >>> import causalpy as cp
+    >>> sc = cp.load_data("sc")
+    >>> treatment_time = 70
+    >>> seed = 42
+    >>> result = cp.pymc_experiments.PrePostFit(
+    ...     sc,
+    ...     treatment_time,
+    ...     formula="actual ~ 0 + a + b + c + d + e + f + g",
+    ...     model=cp.pymc_models.WeightedSumFitter(
+    ...         sample_kwargs={
+    ...             "draws": 2000,
+    ...             "target_accept": 0.95,
+    ...             "random_seed": seed,
+    ...             "progressbar": False
+    ...         }
+    ...     ),
+    ... )
+    >>> result.summary() # doctest: +NUMBER
+    ==================================Pre-Post Fit==================================
+    Formula: actual ~ 0 + a + b + c + d + e + f + g
+    Model coefficients:
+    a                             0.3, 94% HDI [0.3, 0.3]
+    b                             0.0, 94% HDI [0.0, 0.0]
+    c                             0.3, 94% HDI [0.2, 0.3]
+    d                             0.0, 94% HDI [0.0, 0.1]
+    e                             0.0, 94% HDI [0.0, 0.0]
+    f                             0.1, 94% HDI [0.1, 0.2]
+    g                             0.0, 94% HDI [0.0, 0.0]
+    sigma                         0.2, 94% HDI [0.2, 0.3]
+    """
+
+    def __init__(
+        self,
+        data: pd.DataFrame,
+        treatment_time: Union[int, float, pd.Timestamp],
+        formula: str,
+        model=None,
+        **kwargs,
+    ) -> None:
+        super().__init__(model=model, **kwargs)
+        self._input_validation(data, treatment_time)
+
+        self.treatment_time = treatment_time
+        # set experiment type - usually done in subclasses
+        self.expt_type = "Pre-Post Fit"
+        # split data in to pre and post intervention
+        self.datapre = data[data.index <= self.treatment_time]
+        self.datapost = data[data.index > self.treatment_time]
+
+        self.formula = formula
+
+        # set things up with pre-intervention data
+        y, X = dmatrices(formula, self.datapre)
+        self.outcome_variable_name = y.design_info.column_names[0]
+        self._y_design_info = y.design_info
+        self._x_design_info = X.design_info
+        self.labels = X.design_info.column_names
+        self.pre_y, self.pre_X = np.asarray(y), np.asarray(X)
+        # process post-intervention data
+        (new_y, new_x) = build_design_matrices(
+            [self._y_design_info, self._x_design_info], self.datapost
+        )
+        self.post_X = np.asarray(new_x)
+        self.post_y = np.asarray(new_y)
+
+        # DEVIATION FROM SKL EXPERIMENT CODE =============================
+        # fit the model to the observed (pre-intervention) data
+        COORDS = {"coeffs": self.labels, "obs_indx": np.arange(self.pre_X.shape[0])}
+        self.model.fit(X=self.pre_X, y=self.pre_y, coords=COORDS)
+        # ================================================================
+
+        # score the goodness of fit to the pre-intervention data
+        self.score = self.model.score(X=self.pre_X, y=self.pre_y)
+
+        # get the model predictions of the observed (pre-intervention) data
+        self.pre_pred = self.model.predict(X=self.pre_X)
+
+        # calculate the counterfactual
+        self.post_pred = self.model.predict(X=self.post_X)
+
+        # causal impact pre (ie the residuals of the model fit to observed)
+        pre_data = xr.DataArray(self.pre_y[:, 0], dims=["obs_ind"])
+        self.pre_impact = (
+            pre_data - self.pre_pred["posterior_predictive"].mu
+        ).transpose(..., "obs_ind")
+
+        # causal impact post (ie the residuals of the model fit to observed)
+        post_data = xr.DataArray(self.post_y[:, 0], dims=["obs_ind"])
+        self.post_impact = (
+            post_data - self.post_pred["posterior_predictive"].mu
+        ).transpose(..., "obs_ind")
+
+        # cumulative impact post
+        self.post_impact_cumulative = self.post_impact.cumsum(dim="obs_ind")
+
+    def _input_validation(self, data, treatment_time):
+        """Validate the input data and model formula for correctness"""
+        if isinstance(data.index, pd.DatetimeIndex) and not isinstance(
+            treatment_time, pd.Timestamp
+        ):
+            raise BadIndexException(
+                "If data.index is DatetimeIndex, treatment_time must be pd.Timestamp."
+            )
+        if not isinstance(data.index, pd.DatetimeIndex) and isinstance(
+            treatment_time, pd.Timestamp
+        ):
+            raise BadIndexException(
+                "If data.index is not DatetimeIndex, treatment_time must be pd.Timestamp."  # noqa: E501
+            )
+
+    def plot(self, counterfactual_label="Counterfactual", **kwargs):
+        """
+        Plot the results
+        """
+        fig, ax = plt.subplots(3, 1, sharex=True, figsize=(7, 8))
+
+        # TOP PLOT --------------------------------------------------
+        # pre-intervention period
+        h_line, h_patch = plot_xY(
+            self.datapre.index,
+            self.pre_pred["posterior_predictive"].mu,
+            ax=ax[0],
+            plot_hdi_kwargs={"color": "C0"},
+        )
+        handles = [(h_line, h_patch)]
+        labels = ["Pre-intervention period"]
+
+        (h,) = ax[0].plot(self.datapre.index, self.pre_y, "k.", label="Observations")
+        handles.append(h)
+        labels.append("Observations")
+
+        # post intervention period
+        h_line, h_patch = plot_xY(
+            self.datapost.index,
+            self.post_pred["posterior_predictive"].mu,
+            ax=ax[0],
+            plot_hdi_kwargs={"color": "C1"},
+        )
+        handles.append((h_line, h_patch))
+        labels.append(counterfactual_label)
+
+        ax[0].plot(self.datapost.index, self.post_y, "k.")
+        # Shaded causal effect
+        h = ax[0].fill_between(
+            self.datapost.index,
+            y1=az.extract(
+                self.post_pred, group="posterior_predictive", var_names="mu"
+            ).mean("sample"),
+            y2=np.squeeze(self.post_y),
+            color="C0",
+            alpha=0.25,
+        )
+        handles.append(h)
+        labels.append("Causal impact")
+
+        ax[0].set(
+            title=f"""
+            Pre-intervention Bayesian $R^2$: {self.score.r2:.3f}
+            (std = {self.score.r2_std:.3f})
+            """
+        )
+
+        # MIDDLE PLOT -----------------------------------------------
+        plot_xY(
+            self.datapre.index,
+            self.pre_impact,
+            ax=ax[1],
+            plot_hdi_kwargs={"color": "C0"},
+        )
+        plot_xY(
+            self.datapost.index,
+            self.post_impact,
+            ax=ax[1],
+            plot_hdi_kwargs={"color": "C1"},
+        )
+        ax[1].axhline(y=0, c="k")
+        ax[1].fill_between(
+            self.datapost.index,
+            y1=self.post_impact.mean(["chain", "draw"]),
+            color="C0",
+            alpha=0.25,
+            label="Causal impact",
+        )
+        ax[1].set(title="Causal Impact")
+
+        # BOTTOM PLOT -----------------------------------------------
+        ax[2].set(title="Cumulative Causal Impact")
+        plot_xY(
+            self.datapost.index,
+            self.post_impact_cumulative,
+            ax=ax[2],
+            plot_hdi_kwargs={"color": "C1"},
+        )
+        ax[2].axhline(y=0, c="k")
+
+        # Intervention line
+        for i in [0, 1, 2]:
+            ax[i].axvline(
+                x=self.treatment_time,
+                ls="-",
+                lw=3,
+                color="r",
+            )
+
+        ax[0].legend(
+            handles=(h_tuple for h_tuple in handles),
+            labels=labels,
+            fontsize=LEGEND_FONT_SIZE,
+        )
+
+        return (fig, ax)
+
+    def summary(self) -> None:
+        """
+        Print text output summarising the results
+        """
+
+        print(f"{self.expt_type:=^80}")
+        print(f"Formula: {self.formula}")
+        # TODO: extra experiment specific outputs here
+        self.print_coefficients()
+
+
+class InterruptedTimeSeries(PrePostFit):
+    """
+    A wrapper around PrePostFit class
+
+    :param data:
+        A pandas dataframe
+    :param treatment_time:
+        The time when treatment occured, should be in reference to the data index
+    :param formula:
+        A statistical model formula
+    :param model:
+        A PyMC model
+
+    Example
+    --------
+    >>> import causalpy as cp
+    >>> df = (
+    ...     cp.load_data("its")
+    ...     .assign(date=lambda x: pd.to_datetime(x["date"]))
+    ...     .set_index("date")
+    ... )
+    >>> treatment_time = pd.to_datetime("2017-01-01")
+    >>> seed = 42
+    >>> result = cp.pymc_experiments.InterruptedTimeSeries(
+    ...     df,
+    ...     treatment_time,
+    ...     formula="y ~ 1 + t + C(month)",
+    ...     model=cp.pymc_models.LinearRegression(
+    ...         sample_kwargs={
+    ...             "target_accept": 0.95,
+    ...             "random_seed": seed,
+    ...             "progressbar": False,
+    ...         }
+    ...     )
+    ... )
+    """
+
+    expt_type = "Interrupted Time Series"
+
+
+class SyntheticControl(PrePostFit):
+    """A wrapper around the PrePostFit class
+
+    :param data:
+        A pandas dataframe
+    :param treatment_time:
+        The time when treatment occured, should be in reference to the data index
+    :param formula:
+        A statistical model formula
+    :param model:
+        A PyMC model
+
+    Example
+    --------
+    >>> import causalpy as cp
+    >>> df = cp.load_data("sc")
+    >>> treatment_time = 70
+    >>> seed = 42
+    >>> result = cp.pymc_experiments.SyntheticControl(
+    ...     df,
+    ...     treatment_time,
+    ...     formula="actual ~ 0 + a + b + c + d + e + f + g",
+    ...     model=cp.pymc_models.WeightedSumFitter(
+    ...         sample_kwargs={
+    ...             "target_accept": 0.95,
+    ...             "random_seed": seed,
+    ...             "progressbar": False,
+    ...         }
+    ...     ),
+    ... )
+    """
+
+    expt_type = "Synthetic Control"
+
+    def plot(self, plot_predictors=False, **kwargs):
+        """Plot the results"""
+        fig, ax = super().plot(counterfactual_label="Synthetic control", **kwargs)
+        if plot_predictors:
+            # plot control units as well
+            ax[0].plot(self.datapre.index, self.pre_X, "-", c=[0.8, 0.8, 0.8], zorder=1)
+            ax[0].plot(
+                self.datapost.index, self.post_X, "-", c=[0.8, 0.8, 0.8], zorder=1
+            )
+        return (fig, ax)
